requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
django-rq==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ffmpeg-python==0.2.0
flake8==7.3.0
future==1.0.0
gunicorn==23.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.9.0
python-dotenv==1.1.1
redis==6.2.0
rq==2.4.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
